Stop the flashing colours
I changed the code so that a specific colour is chosen for each instance of "Obstacle" and so that the colours don't constantly change during the game. I also made an if statement so that none of the instances of "Obstacle" are the same colour as the instances of "Special".